pyproject.toml
[tool.poetry]
name = "users-app-api"
version = "0.1.0"
description = ""
authors = ["jonlarale <jonlarale@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask-restx = "^1.2.0"
flask-bcrypt = "^1.0.1"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
